requirements.txt
jsonpickle
mysql-connector-python
pony
pytest
selenium
pymysql
lxml-4.6.3-cp39-cp39-win_amd64.whl
allure-pytest